Fixed issue preventing lavaca from building
Lavaca has a dependancy on jquery, however it expects that there is an
index.js file.  This is no longer present.

According to Lavaca v2.3.3 the dependency is changed to expect a jquery.js
file, which turns out to be present.

I preferred to apply this fix rather than performing the update to v2.3.3.